Convert project to class.
